add analytics from vercel fix icon folders impreovement
