+++ apps/native/tsconfig.json
@@ -5,7 +5,8 @@
       "@screens/*": ["./src/screens/*"],
       "@templates/*": ["./src/templates/*"],
       "@env": ["../../node_modules/react-native-dotenv"],
-      "@dotenv": ["./env"]
+      "@dotenv": ["./env"],
+      "types/*": ["../../packages/types/src/*"]
     },
     "jsx": "react-native"
   },
