+++ apps/web/.gitignore
@@ -42,3 +42,5 @@ Thumbs.db
 /build
 
 .env
+
+public/exports/*
