+++ apps/web/public/favicon.ico
