+++ apps/web/src/templates/Default.tsx
@@ -4,11 +4,9 @@ import Header from "app/components/Header";
 import Footer from "app/components/Footer";
 import { resume } from "data";
 
-interface PageProps {
-  children?: ReactNode;
-}
+import { DefaultTemplateProps } from "types/components";
 
-export default function DefaultTemplate({ children }: PageProps) {
+export default function DefaultTemplate({ children }: DefaultTemplateProps) {
   return (
     <main className="w-full text-gray bg-[#525659] first-letter:min-h-screen">
       <Header />
