+++ apps/web/tsconfig.json
@@ -8,8 +8,6 @@
     "strict": true,
     "noEmit": true,
     "esModuleInterop": true,
-    "module": "esnext",
-    "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
@@ -17,7 +15,8 @@
     "paths": {
       "@dotenv": ["./env"],
       "@templates/*": ["./src/templates/*"],
-      "@styles/*": ["./src/styles/*"]
+      "@styles/*": ["./src/styles/*"],
+      "types/*": ["../../packages/types/src/*"]
     },
     "plugins": [
       {
