+++ package.json
@@ -31,6 +31,8 @@
     "@apollo/client": "^3.12.9",
     "@expo/vector-icons": "^14.0.4",
     "@sanity/client": "^6.27.2",
+    "@vercel/analytics": "^1.4.1",
+    "@vercel/speed-insights": "^1.1.0",
     "autoprefixer": "^10.4.20",
     "graphql": "^16.10.0",
     "graphql-request": "^7.1.2",
