+++ packages/assets/images/adaptive-icon.png
