+++ packages/assets/images/favicon.png
