+++ packages/assets/images/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "assets",
+  "private": true,
+  "version": "1.0.0",
+  "main": "index.ts",
+  "license": "MIT"
+}
