+++ packages/assets/images/splash-icon.png
