--- packages/assets/index.ts
@@ -1,3 +1,0 @@
-// leave this blank
-// https://github.com/vercel/next.js/issues/12557
-export {};
