--- packages/assets/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "assets",
-  "private": true,
-  "version": "1.0.0",
-  "main": "index.ts",
-  "license": "MIT"
-}
