+++ packages/libs/package.json
@@ -2,6 +2,16 @@
   "name": "libs",
   "private": true,
   "version": "1.0.0",
-  "main": "index.ts",
-  "license": "MIT"
+  "exports": {
+    ".": {
+      "import": "./src/index.ts",
+      "require": "./src/index.ts",
+      "types": "./src/index.ts"
+    },
+    "./apolloClient": {
+      "import": "./src/apolloClient.ts",
+      "require": "./src/apolloClient.ts",
+      "types": "./src/apolloClient.ts"
+    }
+  }
 }
