+++ packages/types/index.ts
@@ -1,0 +1,3 @@
+// leave this blank
+// https://github.com/vercel/next.js/issues/12557
+export {};
