+++ packages/types/src/components/templates.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export interface DefaultTemplateProps {
+  children: ReactNode;
+}
