+++ packages/types/src/ui/image.ts
@@ -1,0 +1,10 @@
+import { ImageProps as RNImageProps } from "react-native";
+
+export interface SharedImageProps extends RNImageProps {
+  className?: string;
+  src: string;
+  alt: string;
+  width?: number;
+  height?: number;
+  fill?: boolean; // Allow Next.js "fill" mode
+}
