+++ packages/types/src/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./icon";
+export * from "./image";
+export * from "./view";
